interface: Remove from the repo
#### Problem

As part of the upgrade to SDK v3, the cyclical dependency between the
sdk crates and the system interface were too much of an annoyance, so
the interface was migrated back into the SDK with
https://github.com/anza-xyz/solana-sdk/pull/264. But the interface code
still lives in this repo as well.

#### Summary of changes

Remove the interface crate and all references to it.